requirements.txt
pyparsing==3.2.0
pyre2==0.3.6
regex==2024.11.6
